[rocjitsu] Improve wmma lowering and use s_endpgm for skipped kernels
rocjitsu hooks track a skipped kernel and error out if the kernel was skipped.
hotswap path however has no runtime hooks to tell it to stop if the kernel
is skipped, so instead just put s_endpgm. Doing s_trap just hangs the gpu for me.

+++ emulation/rocjitsu/lib/rocjitsu/src/rocjitsu/analysis/indirect_branch_discovery.cpp
@@ -1696,7 +1696,8 @@ void recover_vector_lane_stashed_pcs(AnalysisContext &ctx,
       continue;
     }
 
-    if (facts.swappc_ssrc && *facts.swappc_ssrc + 1 < read_halves.size()) {
+    if (facts.swappc_ssrc &&
+        static_cast<size_t>(*facts.swappc_ssrc + 1) < read_halves.size()) {
       const uint16_t pair_lo = *facts.swappc_ssrc;
       const auto &lo = read_halves[pair_lo];
       const auto &hi = read_halves[pair_lo + 1];
